+++ packages/WidgetFormItem.vue
@@ -8,7 +8,8 @@
                :is="getComponent(item.type, item.component)"
                v-model="form[item.prop]"
                v-bind="Object.assign(deepClone(item), params, { size:item.size || 'small' })"
-               :placeholder="item.placeholder || getPlaceholder(item)">
+               :placeholder="item.placeholder || getPlaceholder(item)"
+               :dic="item.dicData">
       <span v-if="params.html"
             v-html="params.html"></span>
     </component>